remove lazyvim, sketchybar notch updates, yabai + borders

+++ fish/.config/fish/functions/yw.fish
@@ -1,10 +1,15 @@
 function yw
+    if not type -q jq
+        echo "$(set_color yellow)yw$(set_color normal) requires the 'jq' command. Please install jq then try again."
+        return 1
+    end
+
     set tmpDir (string join '' "$TMPDIR" "yarn-workspace")
 
     if test "$argv" = --clear-cache
         rm -r $tmpDir
         echo "Cache deleted."
-        return 1
+        return
     end
 
     if not test -d $tmpDir
@@ -16,28 +21,27 @@ function yw
     set cache "$tmpDir/$dirName.txt"
 
     if test -e $cache
-        set list (bat $cache | string join "\n")
+        set workspaces (cat $cache)
     else
         set isYarnV1 (yarn --version | string match '1**')
         if string length -q "$isYarnV1"
-            set list (yarn workspaces --silent info --json | jq -r '[keys][0] []')
+            # jq command turns the Yarn Classic format to match Berry's
+            set workspaces (yarn workspaces --silent info --json | jq -r "to_entries[] | {name: .key, location: .value.location} | tostring")
         else
             set workspaces (yarn workspaces list --json)
-            set list (jq '.name' -r workspaces)
         end
-        set list (string join "\n" $list)
-        echo -e $list >>$cache
+        printf '%s\n' $workspaces >>$cache
     end
+    set list (echo -e $workspaces | jq '.name' -r)
 
-    set workspace (echo -e $list | fzf)
+    set workspace (printf '%s\n' $list | fzf)
     if not string length -q "$workspace"
         return
     end
-    function stripQuotes
-        sed 's/"//g' $argv
-    end
-    set manifest "$(printf '%s\n' $workspaces | rg \"$workspace\" | jq '.location' | stripQuotes)/package.json"
+
+    set manifest "$(printf '%s\n' $workspaces | rg \"$workspace\" | jq '.location' | sed 's/"//g')/package.json"
     set cmd (jq -r '.scripts | keys | .[]' $manifest | fzf)
+
     if not string length -q "$cmd"
         commandline "yarn workspace $workspace "
         return
@@ -45,13 +49,4 @@ function yw
 
     commandline "yarn workspace $workspace $cmd"
     commandline -f execute
-    # if string length -q "$argv"
-    #     commandline "yarn workspace $workspace $argv"
-    #     # yarn workspace $workspace $argv
-    # else
-    #     commandline "yarn workspace $workspace "
-    # end
-    # set_color --dim
-    # echo "yarn workspace $workspace $argv"
-    # set_color normal
 end
